PartyPokemon abstract class
just moving constants for now
