+++ src/gui/view/EvolutionView.java
@@ -15,6 +15,7 @@ import message.MessageUpdate;
 import message.MessageUpdateType;
 import message.Messages;
 import battle.ActivePokemon;
+import pokemon.PartyPokemon;
 import pokemon.PokemonInfo;
 import pokemon.Stat;
 import pokemon.evolution.BaseEvolution;
@@ -186,7 +187,7 @@ class EvolutionView extends View {
         FontMetrics.setFont(g, 30);
         g.setColor(Color.BLACK);
 
-        String preIndex = isEgg ? ActivePokemon.SPRITE_EGG_IMAGE_NAME : preEvolution.getImageName(evolvingPokemon.isShiny());
+        String preIndex = isEgg ? PartyPokemon.SPRITE_EGG_IMAGE_NAME : preEvolution.getImageName(evolvingPokemon.isShiny());
         String postIndex = isEgg ? preEvolution.getImageName(evolvingPokemon.isShiny()) : postEvolution.getImageName(evolvingPokemon.isShiny());
 
         BufferedImage currEvolution = pokemonTiles.getTile(preIndex);
